backend/requirements.txt
apispec==6.6.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
ipykernel==6.29.5
ipython==8.13.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe<3.0.0
matplotlib-inline==0.1.7
mistune==3.0.2
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==3.0.6
python-dotenv==0.19.1
authlib>=1.2.0
requests==2.26.0